requirements.txt
# Enhanced KPI Dashboard System Requirements
# ============================================

# Core Dependencies
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
python-dateutil>=2.8.2

# Excel Processing
openpyxl>=3.1.2
xlsxwriter>=3.1.0
xlrd>=2.0.1

# Document Processing
python-docx>=0.8.11
PyPDF2>=3.0.0
pdfplumber>=0.9.0
markdown>=3.4.0

# Visualization
plotly>=5.17.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Analytics & ML
scikit-learn>=1.3.0
scipy>=1.11.0

# Database (optional)
pymongo>=4.5.0
sqlalchemy>=2.0.0

# Web & API
requests>=2.31.0
beautifulsoup4>=4.12.0

# Utilities
python-dotenv>=1.0.0
click>=8.1.0
tqdm>=4.66.0

# Data Validation
pydantic>=2.0.0
jsonschema>=4.19.0

# Testing (optional)
pytest>=7.4.0
pytest-cov>=4.1.0

# Development (optional)
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0